refact: combine PdfState controllers, rename views
